Drill Tower
Mockup for drill tower
Press space to spawn a drill tower
Only works for mining the north grid
Works is a bit of a stretch
If there is a drill in the right position, it just hides the visibility of
the corresponding tile in the north grid
Doesn't actually break the tile in any way, just hides it

I have no idea how to structure this
The drill doesn't know about the grids
The main script knows about everything, but is it right to have it handle
everything?
Maybe I need a drill manager script in the main scene that just handles
all the drills?