Build Drill From Button
Build drill button
To add more towers, I can add more buttons to the build menu scene and
then change the signal so it sends info about what tower to build

Positioning the UI feels awkward. I'm not sure how to attach it to the
side of the screen or something
It might also be the fault of how I made the camera be centered on (0,0)
instead of the default camera position
Can I make a UI camera that draws over top of the game scene to fix this?