Enemy Spawning
Camps spawn an enemy every 3s
Enemies run for (0,0) (where the castle will be)

Creating the parent for the enemy feels weird. Is there a better way I
should be doing it?

Bug
If holding a tile while it gets drilled, it gets hidden forever
Low priority bc I'm not sure if you'll be able to move the tiles in the
other grids. So the real fix to this bug is to not let the player drag in
the extra grids