New Tile Art
Art is better, but jarring. The tiles don't mesh great together. The
background colours of the tiles should be more similar I think
Contrast is low on the camp
It needs a new sprite anyway

Added fields
Mostly for the player area to make it different than the resource grids
But resource grids can have fields too. Might change this later

Fixed Camp Bug
Camps wouldn't spawn the right amount of enemies
Moving the tiles when a tile is drilled would reset the camp timer so it
would never finish if synced up with the drills