Cannon
Made a cannon tower
When an enemy enters its range, it fires a cannonball
When the cannonball hits an enemy, the enemy dies
Cannon has no reload time/infinite fire rate at the moment. It also only
fires right. But you can park it right over top of the castle and it will
defend it forever