Rotating Towers
Rotate towers in the direction you drag them when you swap them
If you click on a tile to pick it up and move your mouse right, the tower
will face right
I can make it obey rotatation based on the tile picked up by the mouse or
the tile that gets moved out of the way. Not sure which I like more
I might make it so right click rotates towers so you can rotate them in
place. Or I could press R over them to cycle their rotations

Cannons shoot in the direction they are facing
Drills do not care at the moment

spawning
Tile handles the spawning instead of grid so tile can have a reference to
the tower that is on it