Camps Cast Spells
When the camp reaches the middle of the grid, something happens
Currently, it tries to swap a tower with a random neighbour
The camp will pick a position (probably in its row where a cannon would
be). If there is a tower there, it will swap with a random neighbour. If
there is no tower, nothing should happen
So the player should see the camp get into position and move the tower out
of the way

Aiming doesn't work
The intention is to have a camp pick a point near it, but for testing
right now, all camps just select (2,2)

Other Spells
random swap is easy to implement
Some form of pushing might make more sense and be easier for players to
understand.
Stunning the hit tower might also work