Camp Artillery Works
Camps can figure out where to fire their artillery/spells/whatever I'm
calling it. They use magic numbers to figure out what grid they're in and
where the center grid is. So it works now and might break if I move stuff
around later

The camp places an indicator at the target location and then spawns a
projectile to travel the distance. When the projectile gets there, the
camp calls the function in main to do the random swap at the destination
location calculated
The swap takes 3s to get to position and targets the second row/column.
i.e. one behind where the drills would be
Works for North and East grids. Probably works for the other two too.

camp_artillery and camp_artillery_indicator scenes
They're just a node and a sprite
No logic so they could be just the same scene, but with different sprites

Sprites
indicator and projectile use the same sprite. I might change it later. It
kinda makes sense as you're watching to see when they overlap and that's
when the swap happens. It might be good enough

removed some debug messages