Gaining Resources
When you mine, gain resources
When the drill deals damage, it sends a ref to itself along the chain. If
the tile was broken from the damage, the tile will tell the drill what
kind of loot it gets
The drill will create a resource stack scene (either logs or rocks)

Resource Stack
Like a stack of logs that lives on a tile. As opposed to just a counter on the
UI
The idea is to have this exist on a tile and be swapped around to pay for
tower costs
Right now it just rides around with the drill

Future
If the drill has resources in its internal storage, it will try to put
them on the tile behind it
Mining Rules
Drill can't mine if it's full. Needs to dump its contents first
Some more rule comments in can_i_drill() but it's not hooked up yet