Resource Dumping
When the drill has a resource stack in its internal storage, it will try
to dump it on the tile behind it
If the tile behind is empty or the same type, it works
If the drill's storage is full, it won't drill. This happens if it drills
a tree then a mountain. There will be wood behind the drill blocking the
rocks from exiting the drill

Stack on Tiles
When the resource stack gets dumped out, it will now follow that tile
around as it moves
I think I forgot to forbit you from building towers on tiles with a
resource stack...

Combining Resource Stacks
The drill can push like resources behind it
If it has wood in internal buffer and there is wood behind, the wood stack
on the tile will get incremented and the stack on the drill will be set
for deletion. It will move towards the target tile and then queue_free
itself on arrival

Tile + Tile Combining
Not implemented yet. Needs to be drill to tile to combine

Feedback
Can't tell how big stacks are
Can't tell what the drill is doing. is it drilling or not?
Gold is unhandled still. You get 5 from clearing camps