Castle
Made a castle sprite. It sits at the center
When enemies reach the center, they deal damage to the castle. The castle
just prints out dead when it dies

Removed some test UI
I'm still not quiter sure how to make the buttons attach to a side of the
screen. Do they need to be a child of the camera?

Fixed a bug with drilling left and up
Swapped from some if statements to a match and forgot to remove the old if
statements so the drill was swapping the rows twice when it broke a tile