Spooky Tree
A spooky tree is a 3x3 section of all trees. When this happens, they all
turn into spooky trees with different art
This is going to be how to spawn the boss
They also drop 5 wood instead of 3

Detection
Checks whenever you mine a tile. Checks the whole grid not only the row that
got mined
So it shouldn't be able to make a big tree from initial spawn, but as soon
as you mine any tile from that grid, it will turn into a big tree (unless
that tile you mined broke it)
Loops
I could've made loops to check, but it was easy enough to just hard code
the 8-9 checks
There are more magic numbers that only work for a 3x3 group of trees on a
5x5 grid, but oh well

Resource grids
Now only trees and mountains
If a camp spawns, it changes its tile to a field instead
Makes it easier to create a big spooky tree