Boss Movement
When you make a 3x3 of trees, spawn a snake boss
The snake will travel around in a set pattern around the whole map
Appears to work for any spawn location

Need to make the boss behaviour next. Dealing damage to you, taking
damage, growing, shrinking, the tail, etc.

Balance
The spooky trees drop more wood, but you need more stone for cannons. That
seems backwards. You're likely to have a surplus of stone when the boss
spawns so you don't need to mine the wood