Cannon Upgrades and Build 1
After being built for 30s, the cannon will upgrade its range automatically
Needed to change the cannonball to fly farther. The code is a bit sloppy.
cache the timer, but then never use it again so I didn't need to do that

Forest:Mountain ratio
Changed from 50:50 to 55:45 so 3x3 trees is a bit more likely

Title
Added some title art

Tutorial
Tweaked it a bit
Clarified you need to build towers on open tiles. I doubt players will
read it and remember, but its there

Fixed a Corner Drill bug
Drills in corners could crash the game
was checking if pos was 144 to tell if they were in a corner, but they can
be at 120 and be in the corner if they're being dragged
I couldn't get it to crash after fixing so it's probably good enough

Build
Probably good enough to build
I could change the drill art. Maybe after I get the itch page setup